backend/requirements.txt
# ─── Core Django ──────────────────────────────────────────────────────────────
Django>=4.2,<5.0
gunicorn>=21.2,<22.0

# ─── Django REST framework ────────────────────────────────────────────────────
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.3,<5.0
django-filter>=24.0,<25.0
drf-spectacular>=0.27,<1.0

# ─── Database ─────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9,<3.0

# ─── Async task queue ─────────────────────────────────────────────────────────
celery[redis]>=5.3,<6.0
redis>=5.0,<6.0

# ─── File handling ────────────────────────────────────────────────────────────
Pillow>=10.0,<11.0

# ─── Static file serving (Gunicorn-compatible) ───────────────────────────────
whitenoise>=6.6,<7.0

# ─── Config / Environment ─────────────────────────────────────────────────────
python-dotenv>=1.0,<2.0
